+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Changelog
  * added `PuliEvents::PRE_BUILD_REPOSITORY` and `PuliEvents::POST_BUILD_REPOSITORY`
  * added `PackageManager::renamePackage()`
  * removed default servers
+ * renamed `AssetMapping::getPublicPath()` to `getServerPath()`
+ * renamed `InstallationParams::getPublicPath()` to `getServerPath()`
 
 * 1.0.0-beta4 (2015-04-13)